Beta release for the fabric port of Mekanism
* remove use of Lazy to account for mods drawing non-lines in RenderHighlightEvent (#8504, #8376)

* Upload of current changes for the fabric port

* Some Datagen porting

* Update README with alpha status warning

Added warning about alpha status and usage risks.

* [1.20.1 Backport] Backport some QoL improvements for wrench (configurator). (#8298)

* Lot's of bug fixes (mostly model fixes)

* Fix Datagen

* Fix Crash when wearing a Jetpack

* Fix Crash when loading a QIO Drive Array

* Fix jetpack sound

* Lots of Bug fixing
- Fixes Transmitters
and many more

* Fix damage tag bug
Removed old(forge) datagen run files

* Add transaction-aware variants of slot validation predicates
Fixes inserting chemical-convertable items in infuser

* Add datagenerated Files
Bugfixes
- Fluidtanks renders fluid correctly
- Mechanical Pipes renders transported fluid

* Bugfixes
- Transmitters now works with configurable tile entities except energy

* Fix enchantment-based modules, keybinds and module tweaker opening

* Fix HDPEElytra

* Try fixing MekaSuit (just hope it didn't break everything)

* MekanismTools Port

* ModMenu support

* Bugfixes
- Fix some Models

* Comment out Maven section in README (Is currently not true for fabric port)

Commented out the Maven section in the README.

* Migrate forge tags to c and add back missing tags

* Start additions port

* Changes in the config system
- Adding MekanismTiers
- Splitting between config and Item Tier

* Continue additions port

* Continue additions port

* Port Mekanism Additions except Biome & Structure Modifiers
Optimize imports

* Port Mekanism Additions
Optimize imports

* Start Porting Generators
Start Porting Additions Datagen

* Add changes by skipall

* Fix additions datagen & allow running addon datagen & server

* Fix more datagen

* Run additions datagen

* Continue Porting Generators

* Continue Porting Generators
Finish Additions Datagen

* Changes by skipall

* Changes by skippyall

* Fix Crash

* Fix balloon

* datagen

* Balloon position fix

* Fix tools datagen

* tools datagen fix shield

* run tools datagen

* Add 'runFullClient'

* Bugfixes
Rendering changes

* Bugfixes
Rerun datagen

* Meka Tool Render Fix

* Bugfixes

* Fix Build Problems

* Aktualisieren von README.md

Add some known bugs

* Aktualisieren von README.md

Added one known Bug

* Aktualisieren von README.md

Add one known "Bug"

* Modify issue template for Fabric version input

Updated the issue template to replace (Neo)Forge version with Fabric version.

* Aktualisieren von README.md

Add copyright

* Add Mekanism to addon remap classpath

* Beta release for the Fabric Port of Mekanism

Removed a bullet point about crashing when joining a server from the known bugs section.

* Aktualisieren von README.md

Add one known bug

* Update JDK setup and modify publish arguments

---------

+++ src/main/java/mekanism/client/gui/GuiUtils.java
@@ -1,22 +1,10 @@
 package mekanism.client.gui;
 
-import com.google.common.cache.Cache;
-import com.google.common.cache.CacheBuilder;
 import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.BufferBuilder;
-import com.mojang.blaze3d.vertex.BufferUploader;
-import com.mojang.blaze3d.vertex.DefaultVertexFormat;
-import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.blaze3d.vertex.Tesselator;
-import com.mojang.blaze3d.vertex.VertexBuffer;
-import com.mojang.blaze3d.vertex.VertexBuffer.Usage;
+import com.mojang.blaze3d.vertex.*;
 import com.mojang.blaze3d.vertex.VertexFormat.Mode;
 import com.mojang.math.Divisor;
 import it.unimi.dsi.fastutil.ints.IntIterator;
-import java.util.List;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Predicate;
 import mekanism.client.gui.element.GuiElement;
 import mekanism.common.Mekanism;
 import net.minecraft.client.Minecraft;
@@ -35,6 +23,9 @@ import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.joml.Matrix4f;
+
+import java.util.List;
+import java.util.function.Predicate;
 
 public class GuiUtils {
 
@@ -182,7 +173,7 @@ public class GuiUtils {
     }
 
     public static int drawString(GuiGraphics guiGraphics, Font font, Component component, float x, float y, int color, boolean drawShadow) {
-        return guiGraphics.drawString(font, component.getVisualOrderText(), x, y, color, drawShadow);
+        return guiGraphics.drawString(font, component.getVisualOrderText(), (int)x, (int)y, color, drawShadow);
     }
 
     public static void renderItem(GuiGraphics guiGraphics, @NotNull ItemStack stack, int xAxis, int yAxis, float scale, Font font, @Nullable String text, boolean overlay) {
